@@ -1,23 +1,18 @@
-import Vue from 'vue';
-import Router from 'vue-router';
+import {createRouter, createWebHistory} from 'vue-router';
 import Home from './views/Home.vue';
 import Room from './views/Room.vue';
 
-Vue.use(Router);
-
-export default new Router({
-  mode: 'history',
+export default createRouter({
   base: process.env.BASE_URL,
+  history: createWebHistory(),
   routes: [
     {
-      path: '/',
-      name: 'home',
-      component: Home,
+      path: '/room',
+      component: Room,
     },
     {
-      path: '/room',
-      name: 'room',
-      component: Room,
+      path: '/',
+      component: Home,
     },
   ],
 });
