@@ -42,7 +42,7 @@ export default {
         background-size: contain;
         background-position-x: center;
         background-repeat: no-repeat;
-        background-image: url(/img/icon.png);
+        background-image: url('@/../public/img/icon.png');
     }
 
     .app-logo-container .app-logo {
